Fix all typecheck errors across shared, TUI, and web
- Remove unused shared/hooks/useMatrix.ts (TUI and web each have their own)
- Fix Int16Array → Float32Array in AdminView and MainView (encodeWav signature)
- Add kick(), getHomeserverUrl(), getUnreadNotificationCount() to WataService
- Add getApi(), getSyncEngine() accessors to WataClient
- Add kickFromRoom() to MatrixApi
- Accept Uint8Array in sendVoiceMessage() (web audio returns Uint8Array)
- Fix web callers to use WataService methods instead of getClient()
- Remove loglevel from pnpm-lock.yaml

+++ src/tui/views/AdminView.tsx
@@ -169,7 +169,7 @@ export function AdminView({ onBack, currentProfile }: Props) {
       // Generate 0.5 seconds of silence (16kHz PCM)
       const SAMPLE_RATE = 16000;
       const DURATION = 0.5; // seconds
-      const silenceSamples = new Int16Array(Math.floor(SAMPLE_RATE * DURATION)).fill(0);
+      const silenceSamples = new Float32Array(Math.floor(SAMPLE_RATE * DURATION)).fill(0);
 
       // Encode as WAV
       const wavBuffer = encodeWav(silenceSamples, SAMPLE_RATE);
